Extract types in proper package
